add PKCE authorization token tests / address feedback
