+++ integration_test/auth_flow_test.go
@@ -68,6 +68,7 @@ func TestDeviceCodeAuthFlow(t *testing.T) {
 	tfile, err := os.CreateTemp("", "*.integration.json")
 	assert.NoError(t, err)
 	defer os.Remove(tfile.Name())
+	tfile.Close()
 
 	store, err := storage.OpenJsonStore(context.Background(), tfile.Name())
 	assert.NoError(t, err)
